WSSM:Bugfix: Fix stupid truncation of 16 bit _state_to due to 8 bit result
It is really annoying that such an elemenary cast with loss in
resolution is silently done. What does '-Wall -Wextra' do after all? Do
I need an -Wreallyextra?

Did I mention Rust would not allow this kind of stuff unless is explicit?

+++ water_system/WaterSystemSM.cpp
@@ -18,7 +18,7 @@ ulong WaterSystemSM::_timeoutForState(wss_type state)
         system_panic_no_return();
     };
 
-    uint8_t state_to = u16PgmRead(_state_to[state]);
+    uint16_t state_to = u16PgmRead(_state_to[state]);
     // 0 is the default initializer, but for TO we want by default SleepTimeOut
     if (state_to == 0U) {
         state_to = SleepTimeOut;
